tribo-slide-main/MLSTM-FCN-1/pip_requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cached-property==1.5.2
certifi==2021.5.30
cycler==0.11.0
dataclasses==0.8
h5py==2.10.0
html5lib==0.9999999
joblib==0.12.0
Keras==2.0.4
kiwisolver==1.3.1
Markdown==2.2.0
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorflow-gpu==1.2.0
Theano==1.0.5
threadpoolctl==3.1.0
Werkzeug==2.0.3
